build: make client tools warning-clean and enforce -Werror everywhere
Fix the remaining -Wall -Wextra warnings in the bundled clients/examples and
extend -Werror (via the shared ilib_set_warnings helper) to every target:

- iwebreprt: enlarge the date buffers (pretty_start/pretty_stop were only 30
  bytes but the longest 'Weekday DD Month YYYY HH:MM:SS' rendering needs ~37 --
  a real overflow), and initialize hour/lastday/maxval/width.
- ifraggraph: fix a real buffer overflow -- fgets read up to 10240 bytes into a
  1024-byte buffer (now sizeof(text)); initialize the last-point trackers.
- iindex: cast the validated-positive iconW in the unsigned comparison.

Add cmake/IlibWarnings.cmake with ilib_set_warnings(), used by the library,
clients and examples so the whole build is -Werror under ILIB_WERROR (CI).

Also refresh the README note: the code is now ISO C, not K&R.

+++ clients/iindex/iindex.c
@@ -257,7 +257,7 @@ int main ( int argc, char *argv[] )
     /* don't write more text than will fit under the image. */
     len = strlen ( filename );
     ret = ITextWidth ( gc, font, filename, len, &strw );
-    while ( strw > iconW && len > 0 ) {
+    while ( strw > (unsigned int) iconW && len > 0 ) {
       ret = ITextWidth ( gc, font, filename, --len, &strw );
     }
     IDrawString ( image, gc, x, y + iconH + textspace - 2,
